Move `BufferReader` to its own file.
