+++ src/file_reader.cpp
@@ -217,49 +217,4 @@ std::unique_ptr<const Reader> FileReader::sub_reader(offset_t offset, zsize_t si
   return std::unique_ptr<Reader>(new FileReader(source, _offset+offset, size));
 }
 
-
-//BufferReader::BufferReader(std::shared_ptr<Buffer> source)
-//  : source(source) {}
-
-const Buffer BufferReader::get_buffer(offset_t offset, zsize_t size) const
-{
-  return source.sub_buffer(offset, size);
-}
-
-std::unique_ptr<const Reader> BufferReader::sub_reader(offset_t offset, zsize_t size) const
-{
-  auto sub_buff = get_buffer(offset, size);
-  std::unique_ptr<const Reader> sub_read(new BufferReader(sub_buff));
-  return sub_read;
-}
-
-zsize_t BufferReader::size() const
-{
-  return source.size();
-}
-
-offset_t BufferReader::offset() const
-{
-  return offset_t((offset_type)(static_cast<const void*>(source.data(offset_t(0)))));
-}
-
-
-void BufferReader::read(char* dest, offset_t offset, zsize_t size) const {
-  ASSERT(offset.v, <, source.size().v);
-  ASSERT(offset+offset_t(size.v), <=, offset_t(source.size().v));
-  if (! size ) {
-    return;
-  }
-  memcpy(dest, source.data(offset), size.v);
-}
-
-
-char BufferReader::read(offset_t offset) const {
-  ASSERT(offset.v, <, source.size().v);
-  char dest;
-  dest = *source.data(offset);
-  return dest;
-}
-
-
 } // zim
