+++ src/file_reader.h
@@ -49,24 +49,6 @@ class FileReader : public Reader {
     zsize_t _size;
 };
 
-class BufferReader : public Reader {
-  public:
-    BufferReader(const Buffer& source)
-      : source(source) {}
-    virtual ~BufferReader() {};
-
-    zsize_t size() const;
-    offset_t offset() const;
-
-    void read(char* dest, offset_t offset, zsize_t size) const;
-    char read(offset_t offset) const;
-    const Buffer get_buffer(offset_t offset, zsize_t size) const;
-    std::unique_ptr<const Reader> sub_reader(offset_t offset, zsize_t size) const;
-
-  private:
-    const Buffer source;
-};
-
 };
 
 #endif // ZIM_FILE_READER_H_
