+++ test/cluster.cpp
@@ -50,7 +50,7 @@
 #include "../src/cluster.h"
 #include "../src/file_part.h"
 #include "../src/file_compound.h"
-#include "../src/file_reader.h"
+#include "../src/buffer_reader.h"
 #include "../src/writer/cluster.h"
 #include "../src/endian_tools.h"
 #include "../src/config.h"
